snake body draw and grow when it apples

+++ OpenGL-2D/position.cpp
@@ -6,7 +6,7 @@ Position::Position()
 {
 }
 
-Position::Position(int x, int y)
+Position::Position(float x, float y)
 {
 	m_x = x;
 	m_y = y;
@@ -17,22 +17,22 @@ Position::~Position()
 {
 }
 
-void Position::setX(int x)
+void Position::setX(float x)
 {
 	m_x = x;
 }
 
-void Position::setY(int y)
+void Position::setY(float y)
 {
 	m_y = y;
 }
 
-int Position::getX()
+float Position::getX()
 {
 	return m_x;
 }
 
-int Position::getY()
+float Position::getY()
 {
 	return m_y;
 }
